utils: Base for ansible-amazon module
The start & stop as well as the life policies of S3 buckets, were used
by other clients but not in a role. This tends to solve this problem,
making installation centralized and easier to reproduce/modify.

Two scripts are present:
  backup_lifecycle.yml
  start_stop.yml

Which are respectively pushed when variables are defined:
  lifecycle_policies
  start_stop_list

This installs ansible2 in a virtual env, the different packages etc.
The README has been updated accordingly as well.