rds: fix start & stop playbook with conditionals
The playbook needed to rename the 'action' variable which was producing
a warning for ansible: "Found variable using reserved name: action".

An extra condition about the state of the database was also added in
order to remove the possibility of errors when starting an already
started rds instance or the other way around.

Finally, a default region has been setup to avoid issues; note that the
region can be specified per rds instance, rather than globally to gain
in flexibility.